+++ Arduino/Romi-RPi-I2CSlave/motor_hardware.ino
@@ -11,18 +11,18 @@
 
    note: this might be doable directly with the pololu motors object
 */
-int debug_left_motor_power, debug_right_motor_power;
+int16_t debug_left_motor_power, debug_right_motor_power;
 
-int debug_get_left_motor_power() { 
+int16_t debug_get_left_motor_power() { 
   return debug_left_motor_power;
 }
-int debug_get_right_motor_power() {
+int16_t debug_get_right_motor_power() {
   return debug_right_motor_power;
 }
 
-void hw_motors_setspeeds(int left, int right) {
-  int left_motor = left * FLIP_MOTOR_DIR;
-  int right_motor = right * FLIP_MOTOR_DIR;
+void hw_motors_setspeeds(int16_t left, int16_t right) {
+  int16_t left_motor = left * FLIP_MOTOR_DIR;
+  int16_t right_motor = right * FLIP_MOTOR_DIR;
   if ( SWAP_MOTORS ) {
     debug_left_motor_power = left_motor;
     debug_right_motor_power = right_motor;
